+++ e2e/commands/add.e2e.js
@@ -50,6 +50,14 @@ describe('bit add command', function () {
       const bitMap = fs.readFileSync(path.join(helper.localScopePath, '.bit.map.json')).toString();
       expect(bitMap).to.have.string(AUTO_GENERATED_MSG);
     });
+    it('Should not add component to bitmap beacuse test file does not exists', () => {
+      const osComponentName = path.normalize('bar/foo.js');
+      const osFilePathName = path.normalize('bar/foo.spec.js');
+      helper.createComponent('bar', 'foo.js');
+      const addCmd = () => helper.addComponentWithOptions(osComponentName, { t: `${osFilePathName}       ` });
+      expect(addCmd).to.throw(`fatal: the file "${osFilePathName}" was not found`);
+    });
+
     it('Add component from subdir  ../someFile ', () => {
       const barPath = path.join(helper.localScopePath, 'bar/x');
       helper.createComponent('bar', 'foo2.js');
@@ -204,6 +212,26 @@ describe('bit add command', function () {
       expect(files).to.deep.include({ relativePath: 'test/bar/foo2.spec.js', test: true, name: 'foo2.spec.js' });
       expect(files).to.deep.include({ relativePath: 'test/foo2.spec.js', test: true, name: 'foo2.spec.js' });
       expect(files).to.deep.include({ relativePath: 'test/bar/foo.spec.js', test: true, name: 'foo.spec.js' });
+      expect(bitMap).to.have.property('bar/foo');
+    });
+    it('Should add dir files with spec from multiple dsls when test files are placed in same structure but bit add is with glob', () => {
+      helper.createComponent('bar', 'foo.js');
+      helper.createComponent('bar', 'foo2.js');
+      helper.createComponent('bar', 'foo3.js');
+      helper.createComponent('bar', 'foo.spec.js');
+      helper.createComponent('test/bar', 'foo2.spec.js');
+      helper.createComponent('test', 'foo2.spec.js');
+      helper.addComponentWithOptions('bar/*.js', {
+        i: 'bar/foo',
+        m: path.normalize('bar/foo.js'),
+        t: 'test/{PARENT_FOLDER}/{FILE_NAME}.spec.js,bar/foo.spec.js,test/{FILE_NAME}.spec.js'
+      });
+      const bitMap = helper.readBitMap();
+      const files = bitMap['bar/foo'].files;
+      expect(files).to.be.ofSize(6);
+      expect(files).to.deep.include({ relativePath: 'test/bar/foo2.spec.js', test: true, name: 'foo2.spec.js' });
+      expect(files).to.deep.include({ relativePath: 'test/foo2.spec.js', test: true, name: 'foo2.spec.js' });
+      expect(files).to.deep.include({ relativePath: 'bar/foo.spec.js', test: true, name: 'foo.spec.js' });
       expect(bitMap).to.have.property('bar/foo');
     });
     it('Should add dir files with spec from dsl and glob pattern', () => {