+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
+- Add a CLI command to easily get dependencies for a file
 
 ## [0.10.4] - 2017-10-01
 - Add support for writing package.json files