+++ CHANGELOG.md
@@ -6,8 +6,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
+
+- bit add' on missing test file should throw an error
+- prevent test files from becoming new components
 - rename `bit commit` to `bit tag`
 - fix bug when component version is larger than 10 it won't show as staged
+
 
 ## [0.10.8] - 2017-10-01